+++ src/ray.rs
@@ -14,7 +14,7 @@ impl Ray {
         }
     }
 
-    pub fn at(&self, t: f32) -> Point3 {
+    pub fn at(&self, t: f64) -> Point3 {
         self.orig + t * self.dir
     }
